fix:レビューの指摘箇所の修正
