+++ app/circles/[circle_id]/activities/[activity_id]/edit/page.tsx
@@ -3,7 +3,7 @@ import { notFound } from "next/navigation"
 import { getCircleById, getCircles } from "@/actions/circle/fetch-circle"
 import { auth } from "@/auth"
 import { ActivityForm } from "@/components/forms/activity-form"
-import { MetadataSet } from "@/components/tab/tab-title"
+import { MetadataSet } from "@/utils/metadata"
 import { getActivities, getActivityById } from "@/data/activity"
 
 interface Props {
@@ -14,23 +14,6 @@ interface Props {
 }
 export const generateMetadata = ({ params }: Props) =>
   MetadataSet(params.circle_id || "", "活動日程編集")
-
-/*export const generateMetadata = async ({ params }: Props) => {
-  const { circle_id } = params
-  const circle = await getCircleById(circle_id || "")
-
-  if (!circle) {
-    return {
-      title: "サークルが見つかりません。",
-      description: "サークルが見つかりません。",
-    }
-  }
-
-  return {
-    title: circle.name,
-    description: circle.description,
-  }
-}*/
 
 export const generateStaticParams = async () => {
   const circles = await getCircles()
