+++ app/circles/[circle_id]/announcement/[topic_id]/page.tsx
@@ -3,7 +3,7 @@ import { getCircleById, getCircles } from "@/actions/circle/fetch-circle"
 import { getMembershipRequests } from "@/actions/circle/membership-request"
 import { auth } from "@/auth"
 import { CircleDetailPage } from "@/components/layouts/circle-detail-page"
-import { MetadataSet } from "@/components/tab/tab-title"
+import { MetadataSet } from "@/utils/metadata"
 import { getAnnouncementById, getAnnouncements } from "@/data/announcement"
 
 interface Props {
