+++ app/circles/[circle_id]/edit/page.tsx
@@ -4,7 +4,7 @@ import { notFound } from "next/navigation"
 import { getCircleById, getCircles } from "@/actions/circle/fetch-circle"
 import { auth } from "@/auth"
 import { CircleForm } from "@/components/forms/circle-form"
-import { MetadataSet } from "@/components/tab/tab-title"
+import { MetadataSet } from "@/utils/metadata"
 import { getInstructors } from "@/data/circle"
 
 interface Props {
