+++ app/circles/[circle_id]/thread/[topic_id]/edit/page.tsx
@@ -3,7 +3,7 @@ import { notFound } from "next/navigation"
 import { getCircleById, getCircles } from "@/actions/circle/fetch-circle"
 import { auth } from "@/auth"
 import { ThreadForm } from "@/components/forms/thread-form"
-import { MetadataSet } from "@/components/tab/tab-title"
+import { MetadataSet } from "@/utils/metadata"
 import { getThreadById, getThreads } from "@/data/thread"
 
 interface Props {
